Move user-facing buffer storage to Metal module
