+++ lib/mtl/storage_type.jl
@@ -1,63 +1,4 @@
-export SharedStorage, ManagedStorage, PrivateStorage
 export ReadUsage, WriteUsage, ReadWriteUsage
-
-# Metal Has 4 storage types
-# SharedStorage  -> Buffer in Host memory, accessed by the GPU. Requires no sync
-# ManagedStorage -> Mirrored memory buffers in host and GPU. Requires syncing
-# PrivateStorage -> Memory in Device, not accessible by Host.
-# Memoryless -> render pipeline stuff. ignore it (for now)
-
-abstract type StorageMode end
-
-"""
-    struct Metal.SharedStorage <: MTL.StorageMode
-
-Used to indicate that the resource is stored using `MTLStorageModeShared` in memory.
-
-For more information on Metal storage modes, refer to the official Metal documentation.
-
-See also [`Metal.PrivateStorage`](@ref).
-"""
-struct SharedStorage  <: StorageMode end
-
-"""
-    struct Metal.ManagedStorage <: MTL.StorageMode
-
-Used to indicate that the resource is stored using `MTLStorageModeManaged` in memory.
-
-For more information on Metal storage modes, refer to the official Metal documentation.
-
-!!! warning
-    `ManagedStorage` is no longer supported with `MtlArray`s. Instead, use `SharedStorage` or use the Metal api directly from `Metal.MTL`.
-
-See also [`Metal.SharedStorage`](@ref) and [`Metal.PrivateStorage`](@ref).
-"""
-struct ManagedStorage <: StorageMode end
-
-"""
-    struct Metal.PrivateStorage <: MTL.StorageMode
-
-Used to indicate that the resource is stored using `MTLStorageModePrivate` in memory.
-
-For more information on Metal storage modes, refer to the official Metal documentation.
-
-See also [`Metal.SharedStorage`](@ref).
-"""
-struct PrivateStorage <: StorageMode end
-struct Memoryless     <: StorageMode end
-
-# Remove the ".MTL" when printing
-Base.show(io::IO, ::Type{<:PrivateStorage}) = print(io, "Metal.PrivateStorage")
-Base.show(io::IO, ::Type{<:SharedStorage}) = print(io, "Metal.SharedStorage")
-
-"""
-    MTL.CPUStorage
-
-Union type of [`Metal.SharedStorage`](@ref) and [`Metal.ManagedStorage`](@ref) storage modes.
-
-Represents storage modes where the resource is accessible via the CPU.
-"""
-const CPUStorage = Union{SharedStorage,ManagedStorage}
 
 Base.convert(::Type{MTLStorageMode}, ::Type{SharedStorage})     = MTLStorageModeShared
 Base.convert(::Type{MTLStorageMode}, ::Type{ManagedStorage})    = MTLStorageModeManaged
